backend/pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "rudnevo-app"
version = "0.1.0"
description = "FastAPI-based application with PostgreSQL backend"
authors = [
    {name = "terlyne", email = "kopelevichde@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"

[tool.poetry]
name = "rudnevo-app"
version = "0.1.0"
description = "FastAPI-based admin panel application with PostgreSQL backend"
authors = ["terlyne <kopelevichde@gmail.com>"]
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
fastapi = ">=0.115.12,<0.116.0"
sqlalchemy = ">=2.0.41,<3.0.0"
pydantic-settings = ">=2.9.1,<3.0.0"
pydantic = {version = ">=2.11.5,<3.0.0", extras = ["email"]}
passlib = ">=1.7.4,<2.0.0"
python-jose = {version = ">=3.5.0,<4.0.0", extras = ["cryptography"]}
uvicorn = ">=0.34.3,<0.35.0"
asyncpg = ">=0.30.0,<0.31.0"
python-multipart = ">=0.0.20,<0.0.21"
pillow = ">=11.2.1,<12.0.0"
fastapi-mail = ">=1.5.0,<2.0.0"
bcrypt = ">=4.3.0,<5.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
